refactor: created MinimizeBase
