+++ pysource_minimize/_utils.py
@@ -1,0 +1,9 @@
+import sys
+
+
+if sys.version_info >= (3, 8):
+    from ast import unparse
+else:
+    from astunparse import unparse  # type: ignore
+
+__all__ = ("unparse",)
