+++ tests/utils.py
@@ -1,13 +1,13 @@
 from contextlib import contextmanager
 
-import pysource_minimize._minimize
+import pysource_minimize._minimize_base
 
 
 @contextmanager
 def testing_enabled():
-    old_value = pysource_minimize._minimize.TESTING
-    pysource_minimize._minimize.TESTING = True
+    old_value = pysource_minimize._minimize_base.TESTING
+    pysource_minimize._minimize_base.TESTING = True
     try:
         yield
     finally:
-        pysource_minimize._minimize.TESTING = old_value
+        pysource_minimize._minimize_base.TESTING = old_value
